+++ cmd/eliza/input_pending_windows.go
@@ -1,0 +1,7 @@
+//go:build windows
+
+package main
+
+func readPendingTerminalInput() (string, error) {
+	return "", nil
+}